ROC.ID.Digit

Documentation

data Digit

Represents a single decimal digit in the range 0 to 9.

Constructors

D0

D1

D2

D3

D4

D5

D6

D7

D8

D9

fromChar :: Char -> Maybe Digit

Attempts to parse a Digit from a character.

The Char must be a decimal digit in the range 0 to 9.

toChar :: Digit -> Char

Converts a Digit to a decimal digit character.