ROC.ID.Location

Documentation

data Location

A location, encodable within an ROC identification number.

To generate the name of a Location, use the toText function.

To parse a Location, use the fromChar function.

To generate a random Location, use the generate function.

fromChar :: Char -> Maybe Location

Parse the specified uppercase alphabetic character as a Location.

Returns Nothing if the specified character is not an uppercase alphabetic character.

toText :: Language -> Location -> Text

Pretty-print the specified Location.

generate :: MonadRandom m => m Location

Generate a random Location.