ROC.ID.Serial

Documentation

newtype Serial

A 7-digit serial number, as found within an ROC identification number.

A serial number is unique for a gender and location.

To generate a random Serial number, use the randomSerial function.

Constructors

Serial (Vector 7 Digit)

randomSerial :: MonadRandom m => m Serial

Generate a random Serial number.