ROC.ID

Documentation

data Identity

Represents a valid 10-digit ROC national identification number (中華民國身份證號碼) of the form A123456789.

By construction, invalid values are not representable by this type.

An identification number encodes a person's Gender, the Location in which they first registered for an identification card, and a unique Serial number.

Constructors

Identity

Fields

gender :: !Gender

The gender of the person to whom this ID number belongs.

location :: !Location

The location in which the person first registered for an ID card.

nationality :: !Nationality

The nationality of the person to whom this ID number belongs.

serial :: !Serial

The serial number portion of this ID number.

checksum :: Identity -> Digit

Calculate the checksum of the specified Identity.

fromNumber :: IdentityNumber -> Identity

fromText :: Text -> Either FromTextError Identity

Attempt to parse an Identity using the specified Text as input.

The input must be of the form A123456789.

data FromTextError

Constructors

InvalidLength

InvalidChecksum

InvalidChar CharIndex CharSet

newtype CharIndex

Constructors

CharIndex Digit

data CharSet

Constructors

CharSet (NESet Char)

CharRange Char Char

toNumber :: Identity -> IdentityNumber

toText :: Identity -> Text

generate :: MonadRandom m => m Identity

Generate a random Identity.