ROC.ID

Documentation

data Identity

Represents a valid 10-digit ROC national identification number (中華民國身份證號碼) of the form A123456789.

By construction, invalid values are not representable by this type.

An identification number encodes a person's Gender, the Location in which they first registered for an identification card, and a unique Serial number.

Constructors

Identity

Fields

identityGender :: Gender

The gender of the person to whom this ID number belongs.

identityLocation :: Location

The location in which the person first registered for an ID card.

identitySerial :: Serial

The serial number portion of this ID number.

identityChecksum :: Identity -> Digit

Calculate the checksum of the specified Identity.

parseIdentity :: Text -> Either ParseError Identity

Attempt to parse an Identity using the specified Text as input.

The input must be of the form A123456789.

data ParseError

An error produced when parsing an Identity with the parseIdentity function.

Constructors

InvalidLength
The input was either too short or too long.

InvalidGender
The gender portion of the input was invalid.

InvalidLocation
The location portion of the input included non-alphabetic characters.

InvalidSerial
The serial number portion of the input included non-numeric characters.

InvalidChecksum
The computed checksum did not match the checksum portion of the input.

randomIdentity :: MonadRandom m => m Identity

Generate a random Identity.