ROC.ID

Type

data ID

Represents a valid 10-digit ROC (Taiwan) Uniform Identification Number (中華民國統一證號) of the form A123456789.

By construction, invalid identification numbers are not representable by this data type.

To guarantee correctness, an ID value does not store the terminal checksum digit of the identification number it represents.

To compute the checksum digit, use checksumDigit.

Constructors

ID

Fields

c0 :: !Letter

c1 :: !Digit1289

c2 :: !Digit

c3 :: !Digit

c4 :: !Digit

c5 :: !Digit

c6 :: !Digit

c7 :: !Digit

c8 :: !Digit

Construction

fromSymbol :: forall (s :: Symbol). ValidID s => ID

Constructs an ID from a type-level textual symbol.

The symbol must be exactly 10 characters in length and of the form A123456789.

More precisely, the symbol must match the regular expression ^[A-Z][1289][0-9]{8}$, and the resultant ID must have a valid checksum.

Using a symbol that does not satisfy these constraints will result in a compile-time error.

Requirements

>>> :set -XDataKinds
>>> :set -XTypeApplications
>>> import qualified ROC.ID as ID

Usage

>>> ID.fromSymbol @"A123456789"
ID.fromSymbol @"A123456789"

fromText :: Text -> Either FromTextError ID

Attempts to construct an ID from Text.

The input must be exactly 10 characters in length and of the form A123456789.

More precisely, the input must match the regular expression ^[A-Z][1289][0-9]{8}$, and the resultant ID must have a valid checksum.

This function satisfies the following law:

fromText (toText i) == Right i

data FromTextError

Indicates an error that occurred while constructing an ID from Text.

Constructors

TextTooShort
Indicates that the input text is too short.

TextTooLong
Indicates that the input text is too long.

InvalidChar CharIndex CharSet
Indicates that the input text contains a character that is not allowed.

CharIndex specifies the position of the offending character.

CharSet specifies the set of characters allowed at that position.

InvalidChecksum
Indicates that the parsed identification number has an invalid checksum.

newtype CharIndex

Specifies the zero-based position of a character within a string.

Constructors

CharIndex Int

data CharSet

Specifies a set of characters.

Constructors

CharSet (NESet Char)
An explicit set of characters.

CharRange Char Char
An inclusive range of characters.

Conversion

toText :: ID -> Text

Converts an ID to Text.

The output is of the form A123456789.

This function satisfies the following law:

fromText (toText i) == Right i

Validity

checksumDigit :: ID -> Digit

Computes the terminal checksum digit of an ID.

Generation

generate :: MonadRandom m => m ID

Generates a random ID.

Inspection

getGender :: ID -> Gender

Decodes the Gender component of an ID.

getLocation :: ID -> Location

Decodes the Location component of an ID.

getNationality :: ID -> Nationality

Decodes the Nationality component of an ID.

Modification

setGender :: Gender -> ID -> ID

Updates the Gender component of an ID.

setLocation :: Location -> ID -> ID

Updates the Location component of an ID.

setNationality :: Nationality -> ID -> ID

Updates the Nationality component of an ID.